pyproject.toml
[project]
name = "habits-tracker-spa"
version = "0.1.0"
description = ""
authors = [
    {name = "Максим Малеев",email = "mmaleev1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "django (>=5.2.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "pytest-cov (>=6.1.1,<7.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "eventlet (>=0.40.0,<0.41.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "flake8 (>=7.2.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
